[Refactor] flat the folders at `./src/editor/common/`
