+++ src/editor/common/editorConfiguration.ts
