+++ src/editor/common/editorExtension.ts
