+++ src/editor/model/editorModel.ts
@@ -5,8 +5,8 @@ import { URI } from "src/base/common/files/uri";
 import { defaultLog, ILogService } from "src/base/common/logger";
 import { AsyncResult, ok } from "src/base/common/result";
 import { assert } from "src/base/common/utilities/panic";
-import { EditorOptionsType } from "src/editor/common/configuration/editorConfiguration";
-import { IEditorExtension } from "src/editor/common/extension/editorExtension";
+import { EditorOptionsType } from "src/editor/common/editorConfiguration";
+import { IEditorExtension } from "src/editor/common/editorExtension";
 import { IEditorModel } from "src/editor/common/model";
 import { IEditorPosition } from "src/editor/common/position";
 import { ProseEditorState, ProseNode, ProseTransaction } from "src/editor/common/proseMirror";
