+++ src/editor/view/viewPart/editor/editorBase.ts
@@ -2,7 +2,7 @@ import 'src/editor/view/media/editor/editorBase.scss';
 import { ProseEditorState, ProseEditorView } from "src/editor/common/proseMirror";
 import { ViewContext } from "src/editor/view/editorView";
 import { EditorViewProxy, IEditorViewProxy } from "src/editor/view/viewPart/editor/adapter/editorViewProxy";
-import { IEditorExtension } from 'src/editor/common/extension/editorExtension';
+import { IEditorExtension } from 'src/editor/common/editorExtension';
 
 export interface IEditorBase extends IEditorViewProxy {
