+++ src/editor/view/viewPart/editor/richtextEditor.ts
@@ -1,7 +1,7 @@
 import 'src/editor/view/media/editor/richtextEditor.scss';
 import { ViewContext } from "src/editor/view/editorView";
 import { EditorBase } from 'src/editor/view/viewPart/editor/editorBase';
-import { IEditorExtension } from 'src/editor/common/extension/editorExtension';
+import { IEditorExtension } from 'src/editor/common/editorExtension';
 import { ProseEditorState } from 'src/editor/common/proseMirror';
 
 export class RichtextEditor extends EditorBase {
